Progress with authentication

+++ common/common.h
@@ -7,7 +7,7 @@
     int addToEpoll(int epollfd, int fd, uint32_t flags);
     int changeEpoll(int epollfd, int fd, uint32_t flags);
     int writeNonBlock(int fd, char *string);
-    int readNonBlock(int fd, char **buffer, size_t beginSize);
+    ssize_t readNonBlock(int fd, char **buffer, size_t beginSize);
     char *cleanString(char *string);
     #define COMMON_H
 #endif //COMMON_H
